+++ examples/example-split-LD.R
@@ -2,22 +2,17 @@
 
   corr <- readRDS(url("https://www.dropbox.com/s/65u96jf7y32j2mj/spMat.rds?raw=1"))
 
-  THR_R2 <- 0.01
-
-  # You can provide multiple max_size to try
-  # For real data, use e.g. round(seq_log(1000 + ncol(corr) / 30,
-  #                                       5000 + ncol(corr) / 10,
-  #                                       length.out = 6))
-  SEQ <- c(50, 80)
-  (res <- snp_ldsplit(corr, thr_r2 = THR_R2, min_size = 10, max_size = SEQ,
-                      max_K = 50, max_r2 = 0.5, max_cost = Inf))
+  THR_R2 <- 0.02
+  m <- ncol(corr)
+  (SEQ <- round(seq_log(m / 30, m / 5, length.out = 10)))
+  (res <- snp_ldsplit(corr, thr_r2 = THR_R2, min_size = 10, max_size = SEQ))
 
   library(ggplot2)
   # trade-off cost / number of blocks
   qplot(n_block, cost, color = factor(max_size, SEQ), data = res) +
     theme_bw(14) +
     scale_y_log10() +
-    theme(legend.position = c(0.5, 0.82)) +
+    theme(legend.position = "top") +
     labs(x = "Number of blocks", color = "Maximum block size",
          y = "Sum of squared correlations outside blocks")
 
@@ -25,17 +20,16 @@
   qplot(perc_kept, cost, color = factor(max_size, SEQ), data = res) +
     theme_bw(14) +
     # scale_y_log10() +
-    theme(legend.position = c(0.6, 0.82)) +
+    theme(legend.position = "top") +
     labs(x = "Percentage of non-zero values kept", color = "Maximum block size",
          y = "Sum of squared correlations outside blocks")
 
   # trade-off cost / sum of squared sizes
-  res$cost2 <- sapply(res$all_size, function(sizes) sum(sizes^2))
   qplot(cost2, cost, color = factor(max_size, SEQ), data = res) +
     theme_bw(14) +
     scale_y_log10() +
     geom_vline(xintercept = 0)+
-    theme(legend.position = c(0.7, 0.82)) +
+    theme(legend.position = "top") +
     labs(x = "Sum of squared blocks", color = "Maximum block size",
          y = "Sum of squared correlations outside blocks")
 
@@ -43,10 +37,9 @@
   ## Pick one solution and visualize blocks
   library(dplyr)
   all_ind <- res %>%
-    filter(cost < 1) %>%
-    arrange(perc_kept) %>%
-    slice(1) %>%
+    arrange(cost2 * sqrt(5 + cost)) %>%
     print() %>%
+    slice(1) %>%
     pull(all_last)
 
   ## Transform sparse representation into (i,j,x) triplets